+++ example/lib/main.dart
@@ -27,8 +27,10 @@ import 'package:z_components/components/z-estrutura-empresa/z-estrutura-empresa.
 import 'package:z_components_example/main-components.dart';
 import 'package:z_components_example/main-testing.dart';
 import 'package:z_components_example/material_forms_sample.dart';
+import 'package:z_components_example/tela-teste-card.dart';
 
 import 'main-components.dart';
+import 'main_v2.dart';
 
 void main() {
   initializeDateFormatting().then((_) => runApp(MyApp()));
@@ -39,8 +41,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false,
-      home: new MainComponents(),
-      theme: CustomTheme.appTheme,
+      home: new TelaTesteCard(),
     );
   }
 }